pyproject.toml
[project]
name = "biodb"
dynamic = ["version"]
authors = [
    {name = "Scientifc Software Engineering Center at JHU", email = "ssec@jhu.edu"},
]
description = "Spectroscopy for Patient Diagnosis Database (SPaDDa)"
readme = "README.md"
license = {file = "LICENSE"}

requires-python = ">=3.11"
dependencies = [
    "boto3",
    "django",
    "django-crontab",
    "django-decorator-include",
    "django-health-check",
    "django-nested-admin",
    "django-storages[s3]",
    "django-sql-explorer[charts]",
    "gunicorn",
    "kaleido",
    "openpyxl",
    "pandas",
    "plotly",
    "psycopg[binary,pool]",
    "pyyaml",
    "whitenoise",
    "xlsxwriter"
]

[project.optional-dependencies]
dev = [
    "bandit>=1.7",
    "build",
    "factory-boy",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "ruff",
    "setuptools>=61.2",
    "setuptools_scm[toml]>=8.0",
    "tox"
]
docs = [
    "sphinx",
    "sphinx_rtd_theme",
    "sphinx-automodapi",
    "nbsphinx"
]

[project.urls]
homepage = "https://github.com/rispadd/biodb"
documentation = "https://github.com/rispadd/biodb"
repository = "https://github.com/rispadd/biodb"

[build-system]
requires = ["setuptools>=61.2", "setuptools_scm[toml]>=8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["biodb"]
zip-safe = false
include-package-data = true

[tool.setuptools_scm]
fallback_version = "1.1.3"
version_file = "biodb/_version.py"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "biodb.settings.dev"
markers = [
    "allow_aliases: override setiings.EXPLORER_DATA_EXPORTERS_ALLOW_DATA_FILE_ALIAS (where applicable)",
    "media_root: override settings.MEDIA_ROOT (where applicable)",
    "include_data_files: override settings.EXPLORER_DATA_EXPORTERS_INCLUDE_DATA_FILES (where applicable)",
    "sql: specifiy SQL query (where applicable)",
    "auto_find_previous_visit: override settings.AUTO_FIND_PREVIOUS_VISIT"]

[tool.bandit]
exclude_dirs = ["tests"]

[tool.ruff]
line-length = 120
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "docs",
    "migrations",
    "node_modules",
    "venv",
    "_version.py"
]

ignore = [
]

target-version = "py311"
